fix aftermerge
