+++ YogurtCleaning.Business/Services/Interfaces/ICleaningObjectsService.cs
@@ -4,7 +4,7 @@ namespace YogurtCleaning.Business.Services
 {
     public interface ICleaningObjectsService
     {
-        void UpdateCleaningObject(CleaningObject modelToUpdate, int id);
+        int CreateCleaningObject(CleaningObject cleaningObject, UserValues userValues);
         void UpdateCleaningObject(CleaningObject modelToUpdate, int id, UserValues userValues);
     }
 }
